Working load solution. Socket still needs debugging.

+++ cpp-invoker/src/handler.py
@@ -125,7 +125,8 @@ def schedule_waiting():
 def handle_stop_event(full_path):
     """Delete the element from the running map"""
     process = running_shared_dict.pop_item(full_path)
-    process.terminate()
+    if process:
+        process.terminate()
     info_json = extract_version_info_from_path(full_path)
     running_fm.pop_item_by_key(full_path)
     info_json["state"] = "Stopped"
